pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trustlens"
version = "0.1.0"
description = "Explainable AI-Powered URL Trust Intelligence Engine"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "TrustLens Contributors" },
]
keywords = ["security", "phishing", "ai", "url-analysis", "trust-scoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.20.0",
    "playwright>=1.41.0",
    "httpx>=0.27.0",
    "tldextract>=5.1.0",
    "python-Levenshtein>=0.25.0",
    "structlog>=24.1.0",
    "jinja2>=3.1.0",
    "cryptography>=42.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
]

[project.optional-dependencies]
openai = ["openai>=1.12.0"]
anthropic = ["anthropic>=0.18.0"]
grok = ["openai>=1.12.0"]  # xAI uses OpenAI-compatible SDK
gemini = ["google-generativeai>=0.5.0"]
all-ai = ["openai>=1.12.0", "anthropic>=0.18.0", "google-generativeai>=0.5.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "httpx>=0.27.0",
]

[project.scripts]
trustlens = "trustlens.main:cli"
trustlens-setup = "setup_wizard:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM", "TCH"]
ignore = ["S101", "S104"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
